[lazy] restructure expressions
